# Conflicts:
#	app/src/main/java/com/mobiledevelopment/cryptoyankee/MainActivity.java
#	app/src/main/java/com/mobiledevelopment/cryptoyankee/adapter/CoinAdapter.java
#	app/src/main/java/com/mobiledevelopment/cryptoyankee/db/dao/CoinRepository.java

+++ app/src/main/java/com/mobiledevelopment/cryptoyankee/db/entity/Coin.java
@@ -3,15 +3,16 @@ package com.mobiledevelopment.cryptoyankee.db.entity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 public class Coin {
-    private Integer id;
+    private int id;
     private String name;
-    private Integer priceUsd;
-    private Integer hPriceUsd;
-    private Integer dPriceUsd;
-    private Integer wPriceUsd;
+    private float priceUsd;
+    private float hPriceUsd;
+    private float dPriceUsd;
+    private float wPriceUsd;
 }
